+++ stores/store.go
@@ -3,7 +3,7 @@ package stores
 import "roboncode.com/go-urlshortener/models"
 
 type Store interface {
-	IncCount() int
+	IncCount() int64
 	Create(code string, url string) (*models.Link, error)
 	Read(code string) (*models.Link, error)
 	List(limit int64, skip int64) []models.Link
